implemented simple version that tests which Sources actually need to be rendered. For testing, it currently cuts off 30 pixels from the border to visualize the result.
The biggest problem seems to be getting the width and height of the Canvas into the BasicViewerState. I am also not sure about sideeffects of this, is this method getVisibleSources() only called for rendering, or also when Sliders are being shown for example. And do we need to do this test also for isSourceVisible() as well?